Starting neighbors structures
I'm trying to set up the grid method because we'll have to figure it out anyway.
Technically the nice way to do it involves hashing indices to get coalesced memory access, etc.
For now it's going to be done super naively; atomicly count the number of particles / grid so
far so that they don't overwrite each other... and then grab adjacent grids (also not nicely coalesced).