So that was sad.
Problems fixed? : Forgot to rezero calculated pressures, gradSums. This sadly made everything 'look correct' because the accumulated pressures were closer to the rest than otherwise.
Which brings us to another problem; choosing initial parameters. The rest density affects dPos, so should
probably be reasonably large (for a small, manageable step size)... but at the moment, with our small
number of neighbors, the calculated pressure was orders of magnitude less than it should be.
Seems solveable by making the 'mass' of each fluid particle larger to correspond with the wanted pressure.