OTL questionably correct fluid sim
Okay, so, first off, I am now quite unconvinced that the fluid sim code is actually correct.
But it looks reasonable. GPU fluid code tries to take the mass into account properly (the serial
code doesn't, but since that looks 'correct' too I'm not sure I want to continue tinkering with that
just now.)
I flipped a sign in the spikykernelgrad calculation, which seems to give better results, but only
because that's what other people did; the actual gradient has the minus sign.